#68 - Moved the encounter menu
