+++ apps/spa-3ncount3r/src/bootstrap.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
 
-import App from './app/app';
+import App from './remote-exports/app/app';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
