+++ apps/spa-3ncount3r/src/remote-entry.ts
@@ -1,1 +1,1 @@
-export { default } from './app/app';
+export { default } from './remote-exports/app/app';
